move FakeStaffViewModel like FakeContributorViewModel
